ajax

+++ models/aseguradorasModel.php
@@ -79,10 +79,12 @@ class AseguradorasModel extends Model{
 
     public function drop($id){
         $query = $this->db->connect();
-        $query->query("DELETE FROM asegurados WHERE aseguradora = '$id'");
-        $query->query("DELETE FROM usuarios WHERE user = '$id'");
-        $query->query("DELETE FROM aseguradoras WHERE cif = '$id'");
-        var_dump($query);
+        if($_SESSION['user'] == 'admin'){
+            $query->query("DELETE FROM asegurados WHERE aseguradora = '$id'");
+            $query->query("DELETE FROM usuarios WHERE user = '$id'");
+            $query->query("DELETE FROM averias WHERE aseguradora = '$id'");
+            $query->query("DELETE FROM aseguradoras WHERE cif = '$id'");
+        }
         return $query;
     }
 }
